REFACTOR + CORRECT: Main, Button 추가 및 Map 이름 변경
